Found six errors from the KU database where two KU numbers were associated with the same field number. I've "corrected" these so that the data summarizing code can run, but I did not verify that the changes to the field numbers are the correct changes. The KU catalog number pairs with duplicated field numbers were:
304148
304149

304156
304157

308814
308815

314816
314817

326536
326537

304782
308302

I simply incremented the field number of the second of
each pair by 1. This is probably the correct fix for the
first five errors, as these were probably caused by
failure to update the field number field during
database entry (hence the consecutive KU numbers).
However, the last pair, separated by thousands of KU
numbers is not as obvious. I incremented the field number
for 308302 by 1, but this may not be correct!