Adding script to merge MSG assemblies.
Adding shell script to use ipyrad to merge the assemblies of the reads
of the 3 MSG plates. Previously steps 1 and 2 (demultiplexing and
filtering) were run on each plate. This script merges the demultiplexed
and filtered reads from all three MSG plates into a single assembly, so
that the remaining steps can be executed across all the samples from the
3 plates.

Also, adding the stdout/stderr from running the added script.