Updating Gekko pairs for ecoevolity analyses.
Only 2 of 8 previous pairs were on islands that where connected during
Pleistocene. Adding Gekko gigante and Gekko mindorensis on Masbate and
Panay. For the latter, using different samples from different locality
on Panay than those used for Negros/Panay pair (this pair is also
updated on this commit to make this work).